InfraRed/InfraRed-main/backend/requirements.txt
# ===== Backend (Ingestion API + Workers) =====
# A 와 B 와 C 모두 같은 이미지를 사용합니다.
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis (Streams 포함)
redis==5.2.0

# PostgreSQL
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# HTTP client (Slack/CTI 등)
httpx==0.27.2

# AWS Bedrock
boto3==1.35.71

# Logging / metrics
structlog==24.4.0
prometheus-client==0.21.0

# Util
python-dotenv==1.0.1
orjson==3.10.11
